Relocation unstructured HA status
