+++ pkg/util/ha_status.go
@@ -1,27 +1,18 @@
 package util
 
 import (
-	"fmt"
 	"os"
 
 	"github.com/open-policy-agent/frameworks/constraint/pkg/apis/templates/v1beta1"
-	"github.com/pkg/errors"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
-func getID() string {
+// GetID returns a unique name for the Gatekeeper pod
+func GetID() string {
 	return os.Getenv("POD_NAME")
 }
 
-func blankStatus(id string, generation int64) map[string]interface{} {
-	return map[string]interface{}{
-		"id":                 id,
-		"observedGeneration": generation,
-	}
-}
-
 func GetCTHAStatus(template *v1beta1.ConstraintTemplate) *v1beta1.ByPodStatus {
-	id := getID()
+	id := GetID()
 	for _, status := range template.Status.ByPod {
 		if status.ID == id {
 			return status
@@ -34,7 +25,7 @@ func GetCTHAStatus(template *v1beta1.ConstraintTemplate) *v1beta1.ByPodStatus {
 }
 
 func SetCTHAStatus(template *v1beta1.ConstraintTemplate, status *v1beta1.ByPodStatus) {
-	id := getID()
+	id := GetID()
 	status.ID = id
 	status.ObservedGeneration = template.GetGeneration()
 	for i, s := range template.Status.ByPod {
@@ -47,7 +38,7 @@ func SetCTHAStatus(template *v1beta1.ConstraintTemplate, status *v1beta1.ByPodSt
 }
 
 func DeleteCTHAStatus(template *v1beta1.ConstraintTemplate) {
-	id := getID()
+	id := GetID()
 	var newStatus []*v1beta1.ByPodStatus
 	for _, status := range template.Status.ByPod {
 		if status.ID == id {
@@ -58,119 +49,3 @@ func DeleteCTHAStatus(template *v1beta1.ConstraintTemplate) {
 	template.Status.ByPod = newStatus
 }
 
-// GetHAStatus gets the value of a pod-specific subfield of status
-func GetHAStatus(obj *unstructured.Unstructured) (map[string]interface{}, error) {
-	id := getID()
-	gen := obj.GetGeneration()
-	statuses, exists, err := unstructured.NestedSlice(obj.Object, "status", "byPod")
-	if err != nil {
-		return nil, errors.Wrap(err, "while getting HA status")
-	}
-	if !exists {
-		return blankStatus(id, gen), nil
-	}
-
-	for _, s := range statuses {
-		status, ok := s.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		curID2, ok := status["id"]
-		if !ok {
-			continue
-		}
-		curID, ok := curID2.(string)
-		if !ok {
-			continue
-		}
-		if id == curID {
-			return status, nil
-		}
-	}
-
-	return blankStatus(id, gen), nil
-}
-
-// SetHAStatus sets the value of a pod-specific subfield of status
-func SetHAStatus(obj *unstructured.Unstructured, status map[string]interface{}) error {
-	id := getID()
-	status["id"] = id
-	status["observedGeneration"] = obj.GetGeneration()
-	statuses, exists, err := unstructured.NestedSlice(obj.Object, "status", "byPod")
-	if err != nil {
-		return errors.Wrap(err, "while setting HA status")
-	}
-	if !exists {
-		if err := unstructured.SetNestedSlice(
-			obj.Object, []interface{}{status}, "status", "byPod"); err != nil {
-			return errors.Wrap(err, "while setting HA status")
-		}
-	}
-
-	for i, s := range statuses {
-		curStatus, ok := s.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		curID2, ok := curStatus["id"]
-		if !ok {
-			continue
-		}
-		curID, ok := curID2.(string)
-		if !ok {
-			continue
-		}
-		if id == curID {
-			statuses[i] = status
-			if err := unstructured.SetNestedSlice(
-				obj.Object, statuses, "status", "byPod"); err != nil {
-				return errors.Wrap(err, "while setting HA status")
-			}
-			return nil
-		}
-	}
-
-	statuses = append(statuses, status)
-	if err := unstructured.SetNestedSlice(
-		obj.Object, statuses, "status", "byPod"); err != nil {
-		return errors.Wrap(err, "while setting HA status")
-	}
-	return nil
-}
-
-func DeleteHAStatus(obj *unstructured.Unstructured) error {
-	id := getID()
-
-	statuses, exists, err := unstructured.NestedSlice(obj.Object, "status", "byPod")
-	if err != nil {
-		return errors.Wrap(err, "while deleting HA status")
-	}
-	if !exists {
-		return nil
-	}
-
-	newStatus := make([]interface{}, 0)
-
-	for i, s := range statuses {
-		curStatus, ok := s.(map[string]interface{})
-		if !ok {
-			return fmt.Errorf("element %d in byPod status is malformed", i)
-		}
-		curID2, ok := curStatus["id"]
-		if !ok {
-			return fmt.Errorf("element %d in byPod status is missing an `id` field", i)
-		}
-		curID, ok := curID2.(string)
-		if !ok {
-			return fmt.Errorf("element %d in byPod status' `id` field is not a string: %v", i, curID2)
-		}
-		if id == curID {
-			continue
-		}
-		newStatus = append(newStatus, s)
-	}
-	if err := unstructured.SetNestedSlice(obj.Object, newStatus, "status", "byPod"); err != nil {
-		return errors.Wrap(err, "while writing deleted byPod status")
-	}
-	return nil
-}
